backend/requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic==2.10.4
pydantic-settings==2.7.0

# Google AI dependencies (compatible versions)
google-generativeai==0.8.3

# Vector database
chromadb==0.5.23

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
